다음 단계 진행 전 패키지 구조 변경
