+++ src/main/java/study/calculate/Calculator.java
@@ -1,6 +1,4 @@
-package study;
-
-import study.operator.Operator;
+package study.calculate;
 
 public class Calculator {
