+++ src/main/java/study/calculate/IntNumber.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculate;
 
 public class IntNumber {
